+++ src/components/SortBox.js
@@ -1,4 +1,6 @@
-// import { useMemo } from 'react'
+import { useMemo, useState } from 'react'
+import AmiiboCard from '../components/AmiiboCard'
+import CardPopup from '../components/CardPopup'
 
 function SortBox({ filterAmiibos }) {
   const optionList = [
@@ -7,41 +9,76 @@ function SortBox({ filterAmiibos }) {
     { idx: '2', name: '업데이트순' },
   ]
 
-  const test = filterAmiibos.map(({ name, image, tail }) => ({
-    name,
-    image,
-    tail,
-  }))
+  const sortedAmiibos = useMemo(
+    () =>
+      filterAmiibos
+        .map(({ name, image, tail }) => ({
+          name,
+          image,
+          tail,
+        }))
+        .sort((a, b) => a.name.localeCompare(b.name)),
+    [filterAmiibos],
+  )
+  const reverseSortedAmiibos = useMemo(
+    () =>
+      filterAmiibos
+        .map(({ name, image, tail }) => ({
+          name,
+          image,
+          tail,
+        }))
+        .sort((a, b) => b.name.localeCompare(a.name)),
+    [filterAmiibos],
+  )
+  const latestSortedAmiibos = useMemo(
+    () =>
+      filterAmiibos
+        .map(({ name, image, tail }) => ({
+          name,
+          image,
+          tail,
+        }))
+        .sort((a, b) => a.tail - b.tail),
+    [filterAmiibos],
+  )
 
-  const sortedAmiibos = () => {
-    test.sort((a, b) => a.name.localeCompare(b.name))
-  }
-  const reverseSortedAmiibos = () => {
-    test.sort((a, b) => b.name.localeCompare(a.name))
-  }
-  const latestSortedAmiibos = () => {
-    test.sort((a, b) => a.tail - b.tail)
-  }
+  const [amiiboList, setAmiiboList] = useState(sortedAmiibos)
 
   const onChange = (e) => {
     const userSelect = e.target.value
     if (userSelect === '0') {
-      console.log(sortedAmiibos())
+      setAmiiboList(sortedAmiibos)
     } else if (userSelect === '1') {
-      console.log(reverseSortedAmiibos())
+      setAmiiboList(reverseSortedAmiibos)
     } else {
-      console.log(latestSortedAmiibos())
+      setAmiiboList(latestSortedAmiibos)
     }
   }
 
+  const [clickedCard, setClickedCard] = useState('')
+
+  const clickedImg = filterAmiibos.find((el) => el.tail === clickedCard)
+  const getClickedCard = (clickedCard) => {
+    setClickedCard(clickedCard)
+  }
+
   return (
-    <select onChange={onChange}>
-      {optionList.map(({ idx, name }) => (
-        <option key={idx} value={idx}>
-          {name}
-        </option>
-      ))}
-    </select>
+    <div>
+      <CardPopup img={clickedImg?.image} />
+      <select onChange={onChange}>
+        {optionList.map(({ idx, name }) => (
+          <option key={idx} value={idx}>
+            {name}
+          </option>
+        ))}
+      </select>
+      <AmiiboCard
+        amiiboList={amiiboList}
+        getClickedCard={getClickedCard}
+        clickedCard={clickedCard}
+      />
+    </div>
   )
 }
