+++ src/routes/Home.js
@@ -1,12 +1,9 @@
 import { useState, useEffect, useMemo } from 'react'
-import AmiiboCard from '../components/AmiiboCard'
-import CardPopup from '../components/CardPopup'
 import SortBox from '../components/SortBox'
 
 function Home() {
   const [amiibo, setAmiibo] = useState([])
   const [loading, setLoading] = useState(true)
-  const [clickedCard, setClickedCard] = useState('')
 
   const getAmiibo = async () => {
     try {
@@ -30,19 +27,6 @@ function Home() {
     () => amiibo.filter((update) => update.release.jp === '2015-07-30'),
     [amiibo],
   )
-  // const filterAmiibos = useMemo(
-  //   () =>
-  //     amiibo
-  //       .filter((update) => update.release.jp === '2015-07-30')
-  //       .map(({ name, image, tail }) => ({ name, image, tail }))
-  //       .sort((a, b) => a.name.localeCompare(b.name)),
-  //   [amiibo],
-  // )
-
-  const clickedImg = filterAmiibos.find((el) => el.tail === clickedCard)
-  const getClickedCard = (clickedCard) => {
-    setClickedCard(clickedCard)
-  }
 
   return (
     <div>
@@ -50,13 +34,7 @@ function Home() {
         'Loading...'
       ) : (
         <div>
-          <CardPopup img={clickedImg?.image} />
           <SortBox filterAmiibos={filterAmiibos} />
-          <AmiiboCard
-            filterAmiibos={filterAmiibos}
-            getClickedCard={getClickedCard}
-            clickedCard={clickedCard}
-          />
         </div>
       )}
     </div>
